requirements.txt
opencv-python
face-recognition
tk
pillow
numpy==1.26.4